Extract Edge into its own file
Move the Edge Data.define out of graph.rb into edge.rb for clearer
module organization. No behavior change.
