+++ lib/dag.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "dag/errors"
+require_relative "dag/edge"
 require_relative "dag/result"
 require_relative "dag/success"
 require_relative "dag/failure"
